Update Paging Source Not Update Item

+++ app/src/main/java/com/avalon/calizer/data/remote/insresponse/ApiResponseUserDetails.kt
@@ -79,7 +79,7 @@ data class ApiResponseUserDetails(
         @SerializedName("city_name")
         val cityName: String,
         @SerializedName("city_id")
-        val cityId: Int,
+        val cityId: Long,
         @SerializedName("contact_phone_number")
         val contactPhoneNumber: String,
         @SerializedName("direct_messaging")
